pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "json-razor"
version = "0.1.0"
description = "Reduces JSON, YAML, and NDJSON volume by collapsing repeated structures while preserving the schema."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = ["PyYAML"]

[project.scripts]
json-razor = "json_razor.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["json_razor*"]
